--testrec# changes
LZ77Optimal usage in block splitter with --testrec#, when # was passed, was removed. It was changed to control maximum failed attempts after last best estimated cost. Still the default is 20. If 0 is passed then it will divide last best estimated cost by 16384 to set the exact maximum number of BSR tries with an additional extending it by 20 when still better result is found after maximum number of BSR is tested.