Update LodePNG to 20191107 @https://github.com/google/zopfli
Update LodePNG to 20191107 as per original while maintaining compatibility with KrzYmod changes. Unfortunately there are some differences in displayed output sizes from previous versions. Not sure why it happens. The final output size was, however, the same on my test palette PNG file (not sure if that will be the same for every file).

LodePNG is not my main project. My work on it here is related to maintaining it as ZopfliPNG - to be compliant with this fork's Zopfli. Changes to original code regarding PNG workflow are usually submitted by the community.